requirements.txt
keyboard
selenium
pyautogui
PyGetWindow
python-dotenv